+++ src/main/java/com/nhsoft/neptune/api/domain/MemberCouponDetailVOModel.java
@@ -22,7 +22,7 @@ public class MemberCouponDetailVOModel extends ApiCloudObject {
     @ApiCloudField(value="券id", required=true)
     private String id;
     @ApiCloudField(value="活动明细ID", required=true)
-    private Integer activityDetailId;
+    private Long activityDetailId;
     @ApiCloudField(value="可用结束时间")
     private String availEnd;
     @ApiCloudField(value="可用开始时间", required=true)
@@ -46,7 +46,7 @@ public class MemberCouponDetailVOModel extends ApiCloudObject {
     @ApiCloudField(value="创建用户中心门店编号")
     private Integer createUserCenterBranchId;
     @ApiCloudField(value="创建用户中心渠道")
-    private Integer customerId;
+    private Long customerId;
     @ApiCloudField(value="创建用户中心单据号")
     private String memo;
     @ApiCloudField(value="券状态(CREATED)")
